fullhistory: fold archiveTip into captiveSource.Tip; de-sleep two tip tests
archiveTip's closure shape existed so the multi-source sampler could hold
it in a source list; with the sampler gone its one production caller was
captiveSource.Tip building the closure only to invoke it. Inline it (the
checkpoint-lag note moves onto the method) and move rootHASGetter next to
its now-sole consumer — tip.go is left holding just tipSource and
sampleTipWithRetry.

TestBackfill_RestartTipUnavailableErrors (né ...Unreachable...) and
TestRun_FirstStartNoTipErrors slipped past the sub-genesis switch and
slept through the production backoff (~4s each) on erroring tips; both
now use tips{0} — one permanent "not ready" poll, same assertions,
~0.1s. Retry exhaustion stays covered by
TestSampleTip_ExhaustedRetriesErrors at a 1ms interval.
